pyproject.toml
[tool.poetry]
name = "pywikicli"
version = "0.1.8"
description = "A CLI tool to interact with MediaWiki via its API"
authors = ["vilosource <viloforge@outlook.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/vilosource/wikibot"
homepage = "https://github.com/vilosource/wikibot"
keywords = ["mediawiki", "wiki", "cli", "bot", "python"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]
packages = [
    { include = "pywikicli" }
]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
requests = "^2.28.0"
PyYAML = "^6.0"
pypandoc = "^1.15"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.10.0"

[tool.poetry.scripts]
pywikicli = "pywikicli.cli:cli"
wikicli = "pywikicli.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
